feat: differentiate full length soundscape and 'extracted' soundscape subsets
We are using 'extracted' sound files, which are subsets of the original full length recording centered around the detection. But we don't know the start and end times of the detection within the original full length recording. Therefore, we can't re-use soundscapes already available in Birdweather that might cover this detection, and we need to systematically publish these 'extracted' soundscape subsets to birdweather.